fix: 모바일 Bottom Sheet 컴포넌트 사라지는 문제 해결
