+++ src/components/symbol-page/SymbolPageClient.tsx
@@ -1,6 +1,7 @@
 'use client';
 
-import { Suspense } from 'react';
+import { Suspense, useState } from 'react';
+import type { ReactNode } from 'react';
 import Link from 'next/link';
 import { ErrorBoundary } from 'react-error-boundary';
 import type { AnalysisResponse } from '@/domain/types';
@@ -9,6 +10,11 @@ import { ChartSkeleton } from '@/components/chart/ChartSkeleton';
 import { ChartErrorFallback } from '@/components/chart/ChartErrorFallback';
 import { TickerAutocomplete } from '@/components/search/TickerAutocomplete';
 import { ChartContent } from '@/components/symbol-page/ChartContent';
+import {
+    MobileAnalysisSheet,
+    SNAP_HALF,
+    type SnapPoint,
+} from '@/components/symbol-page/MobileAnalysisSheet';
 import { SymbolPageProvider } from '@/components/symbol-page/SymbolPageContext';
 import { useTimeframeChange } from '@/components/symbol-page/hooks/useTimeframeChange';
 import { useAssetInfo } from '@/components/symbol-page/hooks/useAssetInfo';
@@ -26,6 +32,8 @@ export function SymbolPageClient({
     initialAnalysisFailed,
     indicatorCount,
 }: SymbolPageClientProps) {
+    const [sheetSnap, setSheetSnap] = useState<SnapPoint>(SNAP_HALF);
+    const [mobileSheetContent, setMobileSheetContent] = useState<ReactNode>(null);
     // Suspense로 인해 ChartContent가 remount될 때 타임프레임 변경 여부를 전달한다.
     // timeframe 변경 횟수를 카운트하여 ChartContent가 타임프레임 변경으로 인해
     // mount된 것인지 초기 mount인지를 구분한다.
@@ -101,10 +109,18 @@ export function SymbolPageClient({
                                 timeframeChangeCount={timeframeChangeCount}
                                 initialAnalysis={initialAnalysis}
                                 initialAnalysisFailed={initialAnalysisFailed}
+                                onMobileSheetContent={setMobileSheetContent}
                             />
                         </Suspense>
                     </ErrorBoundary>
                 </div>
+                {/* Suspense 경계 밖에 렌더링하여 타임프레임 전환 시 바텀시트가 사라지지 않도록 한다 */}
+                <MobileAnalysisSheet
+                    activeSnap={sheetSnap}
+                    onActiveSnapChange={setSheetSnap}
+                >
+                    {mobileSheetContent}
+                </MobileAnalysisSheet>
             </div>
         </SymbolPageProvider>
     );
